Make cleanup automatic
