+++ Sources/Castor/CastPlayer.swift
@@ -82,15 +82,6 @@ public final class CastPlayer: NSObject, ObservableObject {
         _synchronizedActiveTracks.attach(to: remoteMediaClient)
         _synchronizedTargetSeek.attach(to: remoteMediaClient)
     }
-
-    deinit {
-        _synchronizedMediaStatus.detach()
-        _synchronizedShouldPlay.detach()
-        _synchronizedRepeatMode.detach()
-        _synchronizedPlaybackSpeed.detach()
-        _synchronizedActiveTracks.detach()
-        _synchronizedTargetSeek.detach()
-    }
 }
 
 public extension CastPlayer {
