+++ Sources/Castor/CastQueue.swift
@@ -51,11 +51,6 @@ public final class CastQueue: NSObject, ObservableObject {
         _synchronizedItems.attach(to: remoteMediaClient)
         _synchronizedCurrentItemId.attach(to: remoteMediaClient)
     }
-
-    deinit {
-        _synchronizedItems.detach()
-        _synchronizedCurrentItemId.detach()
-    }
 }
 
 public extension CastQueue {
